프로필사진 수정, 삭제 - 23.06.09

+++ JMC/src/main/java/com/kh/jaManChw/mypage/controller/MypageController.java
@@ -37,7 +37,8 @@ public class MypageController {
 		//세션에 저장된 userno 가져오기
 		int userno = (Integer)session.getAttribute("userno");
 		users.setUserno(userno);
-		//users.setUserno(((Integer) session.getAttribute("userno")).intValue());
+		
+		//유저 정보 가져오기 - 메인에서 로그인한 유저 정보 조회
 		Users loginInfo = mypageService.getloginInfo(users);
 		model.addAttribute("loginInfo", loginInfo);
 		
@@ -48,7 +49,11 @@ public class MypageController {
 
 	// 마이페이지 정보수정
 	@GetMapping("/mypage/userInfo")
-	public void userInfoPage() {}
+	public void userInfoPage(Model model,Users users) {
+		Users loginInfo = mypageService.getloginInfo(users);
+		model.addAttribute("loginInfo", loginInfo);
+		
+	}
 	
 	// 유저 정보 수정
 	@PostMapping("/mypage/userInfo")
@@ -94,45 +99,73 @@ public class MypageController {
 	
 	// 프로필사진수정 페이지
 	@GetMapping("/mypage/profile")
-	public void profilePage() {}
+	public void profilePage(HttpSession session,ProfileFile profileFile,Model model) {
+		
+		// 세션에 담긴 userno 가져오기
+		profileFile.setUserno((Integer)session.getAttribute("userno"));
+		
+		//파일 정보 가져오기
+		ProfileFile profile = mypageService.fileInfo(profileFile,model);
+
+		model.addAttribute("profile", profile);
+		logger.info("profile:{}",profile);
+		
+
+				
+	}
 	
 	// 프로필사진 수정
 	@PostMapping("/mypage/profile")
-	public void uploadProfile(MultipartFile file, HttpSession session,ProfileFile profileFile,Model model) {
+	public void uploadProfile(MultipartFile file, HttpSession session,ProfileFile profileFile,Model model,Users users) {
 
 		logger.info("file:{}",file);
 
 		int userno = (Integer)session.getAttribute("userno");
 		profileFile.setUserno(userno);
 		
-		//파일 유무
-		if(!file.isEmpty()) {
+		logger.info("file is empty? {}",!file.isEmpty());
+		
+		int userprofile = mypageService.findcntInfo(profileFile);
+		
+		if(userprofile>0) {
+			
+			// db에 저장된 프로필 삭제
+			mypageService.removeProfile(profileFile);
+			
+			// 프로필 다시 저장
+			mypageService.profileSave(file,session,profileFile);
+		}else {
 			// 프로필 저장
 			mypageService.profileSave(file,session,profileFile);
-			logger.info("file:{}",file);
+			
+			//파일 정보 가져오기
+			ProfileFile profile = mypageService.fileInfo(profileFile,model);
+			
+			model.addAttribute("profile", profile);
+			model.addAttribute("profileStoredName", profile.getProfileStoredName());
+			
+			
+			logger.info("profile",profile);
 		}
 		
-		//파일 정보 가져오기
-		ProfileFile profile = mypageService.fileInfo(profileFile,model);
+//		//파일 유무
+//		if(!file.isEmpty()) {
+//			// 프로필 저장
+//			mypageService.profileSave(file,session,profileFile);
+//			logger.info("file:{}",file);
+//		
+//		}else {
+//			// db에 저장된 프로필 삭제
+//			mypageService.removeProfile(profileFile);
+//			
+//		}
 		
-		model.addAttribute("profile", profile);
-		model.addAttribute("profileStoredName", profile.getProfileStoredName());
-		
-		
-		logger.info("profile",profile);
+		//return;
 		
 		//logger.info("sesion:{}",session.getAttribute("userno"));
 
 	}
 	
-	// 프로필 삭제
-	@PostMapping("/mypage/remove")
-	public void deleteProfile(ProfileFile profileFile) {
-		
-		// db에 저장된 프로필 삭제
-		mypageService.removeProfile(profileFile);
-
-	}
 
 //-----------------------------------------------------------	
 	//친구목록 페이지 페이징
